[netdata] mechanism to detect & signal when network data gets full
This commit adds a new mechanism to detect when Network Data (local
or leader) gets full and signals by invoking a callback which can be
set with a newly added OT API.

New method is added in `NetworkData::Leader` to determine whether
registering local Network Data into the current Leader data can be
successful or if it may lead to Network Data getting full. This is
determined by device emulating the leader behavior and registering
the local Network Data into a clone of the current leader's Network
Data.