[logging] fix format strings (#8446)
This commit fixes the format string in logging functions.

When openthread is included in other projects that use the -Werror
option in their builds, such as Matter, these format string errors
appear. In this case, it can be seen in radio.h that the structure
members are uint32_t but printed with %u.

The size of types depends on the target, there needs to be a casing.