[spinel] save the radio spinel metrics in temporary file
The radio spinel metrics `mRcpTimeoutCount` and `mRcpUnexpectedResetCount`
are used to record the RCP timeout count and RCP unexpected reset count.
The problem is that when the RCP timeout or RCP unexpected reset happens,
the Thread stack will exit. So the value of radio spinel metrics
`mRcpTimeoutCount` and `mRcpUnexpectedResetCount` is always 0.

This commit saves the counters into a temporary file before exiting and
read them from the temporary file when the Thread stack starts.