[mesh-diag] new module and new API to discover network topology
This commit introduces new module `MeshDiag` which provides new API to
discover current network topology.

Information about all routers in mesh are discovered including Router
ID, RLOC16, MAC Extended Address, along with the list of established
links between routers and their link quality. The API also provides
information such as which router is leader, which one is acting as
border router and which router is this device itself or parent of
this device (if we query from a child). The API allows user to
request additional information per router: list of IPv6 addresses per
router, or information about children including child RLOC16, mode,
and link quality.

The implementation uses multiple Network Diagnostic messages to query
routers and then compiles the related information which makes it
compatible with Thread 1.1 devices.

Config `OPENTHREAD_CONFIG_MESH_DIAG_ENABLE` and/or `OT_MESH_DIAG`
CMake option can be used to enable this feature. It is by default
enabled on devices that are configured to act as a border router
`ROUTING_MANAGER` is enabled.

This commit also adds a CLI command `netdiag topology` for this API
along with documentation and examples.